requirements.txt
alembic==1.0.0
anyio==3.5.0
asn1crypto==0.24.0
beautifulsoup4==4.10.0
blinker==1.4
certifi==2021.10.8
cffi==1.14.3
charset-normalizer==2.0.12
click==7.1.2
cryptography==3.4.6
dnspython==2.2.0
dominate==2.5.2
email-validator==1.1.3
favicon==0.7.0
feedwerk==1.0.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
flask-paginate==0.7.1
Flask-Reuploaded==1.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
Pillow
pycparser==2.20
pydantic==1.9.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.19.2
python-editor==1.0.4
requests==2.27.1
six==1.15.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.3.19
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.8
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
